% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialSample.R
\name{mybin}
\alias{mybin}
\title{BinomialSample}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability}
}
\value{
barplot and table of successes
}
\description{
BinomialSample
}
\examples{
mybin(100, 10, 0.7)
}
