% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{sd}

\item{a}{probability to calculate}
}
\value{
probability area
}
\description{
myncurve
}
\examples{
myncurve(10,5,6)
}
