% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{TSS}
\alias{TSS}
\title{TSS Calculation}
\usage{
TSS(df, value)
}
\arguments{
\item{df}{dataframe that calculations will be made off of}

\item{value}{value that calculation will be made off of}
}
\value{
calculated TSS values
}
\description{
TSS Calculation
}
\examples{
TSS(ddt <- read.csv("C:/Users/bella/Documents/School Folders/Fall 2023/Applied Statistical Methods/DATAxls/DDT.csv"), ddt$Height)
}
